requirements.txt
matplotlib
numpy
Pillow
streamlit
pytesseract
opencv-contrib-python-headless
textblob